[BOOK-79] refactor: apis - 요청 DTO값 분리 및 알라딘 외부용 API 분리 (#15)

+++ apis/src/main/kotlin/org/yapp/apis/book/dto/response/BookDetailResponse.kt
@@ -1,6 +1,7 @@
 package org.yapp.apis.book.dto.response
 
 import org.yapp.infra.external.aladin.response.AladinBookDetailResponse
+import java.math.BigDecimal
 
 /**
  * 단일 도서의 상세 정보를 나타내는 DTO.
@@ -16,8 +17,8 @@ data class BookDetailResponse private constructor(
     val isbn: String?,
     val isbn13: String?,
     val itemId: Long?,
-    val priceSales: Double?,
-    val priceStandard: Double?,
+    val priceSales: BigDecimal?,
+    val priceStandard: BigDecimal?,
     val mallType: String?,
     val stockStatus: String?,
     val mileage: Int?,
